extract information for smart agent into a dedicated module
